Move Typography, Uploaders components into directories
