+++ assets/javascripts/kitten/components/typography/sup-title/index.js
