+++ assets/javascripts/kitten/components/uploaders/uploader/index.js
